[DEBT-043-D5] Reduce cli/research/thesis/_commands.py to ≤400 LoC
- Moved _check_thesis_invalidation and _gather_thesis_research_data async
  helpers from _commands.py to _helpers.py
- Updated __init__.py to import from correct module
- Updated test patches to use _shared._get_thesis_file
- _commands.py: 404 → 355 lines
- _helpers.py: 154 → 205 lines
- Quality gates passed: pre-commit, pytest (1052 tests)
- Acceptance criteria: 1/1 complete
